+++ pf1.js
@@ -153,7 +153,7 @@ Hooks.once("setup", function() {
  */
 Hooks.once("ready", async function() {
   // Migrate data
-  const NEEDS_MIGRATION_VERSION = "0.75.13";
+  const NEEDS_MIGRATION_VERSION = "0.76.1";
   let PREVIOUS_MIGRATION_VERSION = game.settings.get("pf1", "systemMigrationVersion");
   if (typeof PREVIOUS_MIGRATION_VERSION === "number") {
     PREVIOUS_MIGRATION_VERSION = PREVIOUS_MIGRATION_VERSION.toString() + ".0";